+++ CotEditor/Sources/DefaultKeys.swift
@@ -36,7 +36,6 @@ extension DefaultKeys {
     static let trimsTrailingWhitespaceOnSave = DefaultKey<Bool>("trimsTrailingWhitespaceOnSave")
     static let trimsWhitespaceOnlyLines = DefaultKey<Bool>("trimsWhitespaceOnlyLines")
     static let documentConflictOption = DefaultKey<Int>("documentConflictOption")
-    static let syncFindPboard = DefaultKey<Bool>("syncFindPboard")
     static let countLineEndingAsChar = DefaultKey<Bool>("countLineEndingAsChar")
     static let autoLinkDetection = DefaultKey<Bool>("autoLinkDetectionKey")
     static let checkSpellingAsType = DefaultKey<Bool>("checkSpellingAsType")