+++ CotEditor/Sources/Script.swift
@@ -10,7 +10,7 @@
  ------------------------------------------------------------------------------
  
  © 2004-2007 nakamuxu
- © 2014-2017 1024jp
+ © 2014-2018 1024jp
  
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
@@ -442,6 +442,9 @@ final class ShellScript: Script {
                     handle.write(chunk)
                 }
                 handle.closeFile()
+                
+                // inPipe must avoid releasing before `writeabilityHandler` is invocated
+                inPipe.fileHandleForWriting.writeabilityHandler = nil
             }
         }