+++ CHANGELOG.md
@@ -30,6 +30,7 @@ develop
 - Append “Option-Command-T” shortcut to “Show/Hide Toolbar” menu item.
 - Restore the last viewed preference pane.
 - Add an input field for window opacity setting.
+- Adjust scroll after toggling line wrap.
 - Tweak acknowledgements window design.
 - Some minor UI improvements and fixes.
 - Update Japanese localization to conform with the modern macOS localization rules.