+++ CHANGELOG.md
@@ -32,6 +32,7 @@ develop
 - Add tooltips to Unicode normalization forms in Text menu.
 - Append “Option-Command-T” shortcut to “Show/Hide Toolbar” menu item.
 - Remove “Color Code Panel” action in “Window” menu (use “Edit Color Code...” action in “Text” menu instead).
+- Remove “share find text with other applications” option.
 - Restore the last viewed preference pane.
 - Add an input field for window opacity setting.
 - Adjust scroll after toggling line wrap.